Interprocess changes to support systems with filesystem-based shared memory
git-svn-id: http://svn.boost.org/svn/boost/trunk@41194 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/interprocess/mapped_region.hpp
@@ -246,7 +246,9 @@ inline mapped_region::mapped_region
          #ifdef max
          #undef max
          #endif
-         if(total_size > std::numeric_limits<std::size_t>::max()){
+
+         if(static_cast<unsigned __int64>(total_size) > 
+            std::numeric_limits<std::size_t>::max()){
             error_info err(size_error);
             throw interprocess_exception(err);
          }
